requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
docstring-parser==0.15
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
gcsfs==2022.8.2
google-api-core==2.10.2
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
hydra-core==1.2.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
jsonargparse==4.15.1
kiwisolver==1.4.4
markdown==3.4.1
matplotlib==3.6.1
multidict==6.0.2
oauthlib==3.2.1
omegaconf==2.2.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeprecate==0.3.2
pygments==2.13.0
pytorch-lightning==1.7.7
pyyaml==6.0
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torchmetrics==0.10.0
torchtext==0.13.1
tqdm==4.64.1
werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
netcdf4==1.6.4
h5netcdf==1.1.0
diffusers